refactor: create a dedicated package `model.theme`
The `AppThemeViewModel.kt` is now refactored in this package.
